avoid compiler warnings: C:\projects\nppftp\src\FTPProfile.cpp(515): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data [C:\projects\nppftp\_build\NppFTP.vcxproj] C:\projects\nppftp\src\FTPProfile.cpp(522): warning C4267: 'argument': conversion from 'size_t' to 'int', possible loss of data [C:\projects\nppftp\_build\NppFTP.vcxproj]

+++ src/FTPProfile.cpp
@@ -512,14 +512,14 @@ Transfer_Mode FTPProfile::GetFileTransferMode(const TCHAR* file) const {
 	LPCTSTR suffix = NULL;
 
 	if (m_asciiTypes.size()) {
-		suffix = PathFindSuffixArray(file, (const TCHAR**)(&m_asciiTypes[0]), m_asciiTypes.size());
+		suffix = PathFindSuffixArray(file, (const TCHAR**)(&m_asciiTypes[0]), static_cast<int>(m_asciiTypes.size()) );
 		if (suffix) {
 			return Mode_ASCII;
 		}
 	}
 
 	if (m_binTypes.size()) {
-		suffix = PathFindSuffixArray(file, (const TCHAR**)(&m_binTypes[0]), m_binTypes.size());
+		suffix = PathFindSuffixArray(file, (const TCHAR**)(&m_binTypes[0]), static_cast<int>(m_binTypes.size()) );
 		if (suffix) {
 			return Mode_Binary;
 		}
